refactor : auth, product, cart 엔티티모델 타입 분리하여 적용
