+++ src/entities/auth/type.ts
@@ -1,0 +1,11 @@
+export interface LoginFormType {
+    email: string;
+    password: string;
+}
+
+export interface SignupFormType {
+    email: string;
+    password: string;
+    nickname: string;
+    isSeller: boolean;
+}
