+++ src/entities/cart/type.ts
@@ -1,3 +1,5 @@
+import { ProductType } from '@/entities/product/type';
+
 export interface CartProps {
     id: number;
     quantity: number;
@@ -8,27 +10,14 @@ export interface CartProps {
     product_id: number;
 }
 
-export interface ProductType {
-    name: string;
-    description: string;
-    origin: string;
-    bean_type: string;
-    price: string;
-    weight: string;
-    stock_quantity: string;
-    product_images: ProductImageType[];
-}
-
-export interface ProductImageType {
-    id: number;
-    image_url: string;
-}
-
 export interface CartFormProps {
-    data: CartProps[];
-    isAllSelected: boolean;
-    selectedItems: number[];
-    totalPrice: number;
-    toggleSelectAll: () => void;
-    toggleSelectItem: (id: number) => void;
+    userId: string;
+    productId: number;
+    option: OptionType;
+}
+
+export interface OptionType {
+    roasting: string;
+    grind: string;
+    quantity: number;
 }
