+++ src/features/cart/model/type.ts
@@ -1,0 +1,10 @@
+import { CartProps } from '@/entities/cart/type';
+
+export interface CartFormProps {
+    data: CartProps[];
+    isAllSelected: boolean;
+    selectedItems: number[];
+    totalPrice: number;
+    toggleSelectAll: () => void;
+    toggleSelectItem: (id: number) => void;
+}
