--- src/features/createProduct/model/type.ts
@@ -1,14 +1,0 @@
-export interface FormType {
-    name: string;
-    description: string;
-    origin: string;
-    bean_type: string;
-    price: string;
-    weight: string;
-    stock_quantity: string;
-}
-
-export interface CreateFormType {
-    form: FormType;
-    images: File[];
-}
