+++ src/features/editCartButton/model/type.ts
@@ -1,7 +1,8 @@
+import { OptionType } from '@/entities/cart/type';
 import { editProductSchema } from '@/features/editCartButton/model/validation';
 import { z } from 'zod';
 
-export interface BuyProductProps {
+export interface EditCartButtonProps {
     price: string;
     id: number;
 }
@@ -9,38 +10,6 @@ export interface BuyProductProps {
 export interface EditCartProps {
     id: number;
     option: OptionType;
-}
-
-interface OptionType {
-    roasting: string;
-    grind: string;
-    quantity: number;
-}
-
-export interface CartProps {
-    id: number;
-    quantity: number;
-    roasting: string;
-    grind: string;
-    products: ProductType;
-    user_id: string;
-    product_id: number;
-}
-
-interface ProductType {
-    name: string;
-    description: string;
-    origin: string;
-    bean_type: string;
-    price: string;
-    weight: string;
-    stock_quantity: string;
-    product_images: ProductImageType[];
-}
-
-interface ProductImageType {
-    id: number;
-    image_url: string;
 }
 
 export type useEditProductFormType = z.infer<typeof editProductSchema>;
