--- src/features/login/model/type.ts
@@ -1,4 +1,0 @@
-export interface LoginType {
-    email: string;
-    password: string;
-}
