--- src/features/recentProducts/model/type.ts
@@ -1,17 +1,0 @@
-export interface ProductsType {
-    id: string;
-    name: string;
-    description: string;
-    origin: string;
-    bean_type: string;
-    price: string;
-    weight: string;
-    stock_quantity: string;
-    product_images: ProductImageType[];
-}
-
-export interface ProductImageType {
-    id: number;
-    image_url: string;
-    product_id: string;
-}
