--- src/pages/ProductDetailPage/model/type.ts
@@ -1,20 +1,0 @@
-export interface ProductDetailProps {
-    data: ProductType;
-}
-
-export interface ProductType {
-    id: number;
-    name: string;
-    description: string;
-    origin: string;
-    bean_type: string;
-    price: string;
-    weight: string;
-    stock_quantity: string;
-    product_images: ProductImageType[];
-}
-
-export interface ProductImageType {
-    id: number;
-    image_url: string;
-}
